requirements.txt
bcrypt==4.1.2
capstone==5.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colored-traceback==0.3.0
cryptography==42.0.5
idna==3.6
intervaltree==3.1.0
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
paramiko==3.4.0
plumbum==1.8.2
psutil==5.9.8
pwntools==4.12.0
pycparser==2.21
pycryptodome==3.20.0
pyelftools==0.30
Pygments==2.17.2
PyNaCl==1.5.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.31.0
ROPGadget==7.4
rpyc==6.0.0
six==1.16.0
sortedcontainers==2.4.0
unicorn==2.0.1.post1
unix-ar==0.2.1
urllib3==2.2.1
zstandard==0.22.0
